Fix lint warning.

+++ src/sc_dmatrix.c
@@ -210,7 +210,7 @@ sc_dmatrix_destroy (sc_dmatrix_t * dmatrix)
 bool
 sc_dmatrix_is_valid (const sc_dmatrix_t * A)
 {
-  return sc_darray_is_valid (A->e[0], A->m * A->n);
+  return sc_darray_is_valid (A->e[0], (size_t) A->m * (size_t) A->n);
 }
 
 bool
